[MINOR] Move HoodieUpgradeDowngradeException to exception package (#1993)
